Move useOnceFullyVisible hook to Misc
